apim_gateway/requirements.txt
openai>=1.82.0
python-dotenv>=1.1.0
rich>=14.0.0
